requirements.txt
google-analytics-data==0.18.5
google-api-core==2.17.0
google-api-python-client==2.103.0
google-auth==2.27.0
google-auth-httplib2==0.1.0
googleapis-common-protos==1.56.4
pandas==1.5.2
python-dateutil==2.8.2
python-dotenv==1.0.1
SQLAlchemy==1.4.45
pymssql==2.2.2
PyMySQL==1.0.2